pyproject.toml
[tool.poetry]
name = "home-work-web-12"
version = "0.1.0"
description = ""
authors = ["pentagon846 <pentagon846@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.110.0"
psycopg2 = "^2.9.9"
sqlalchemy = "^2.0.29"
alembic = "^1.13.1"
fastapi-jwt-auth = "^0.5.0"
python-jose = "^3.3.0"
passlib = "^1.7.4"
python-multipart = "^0.0.9"
uvicorn = {extras = ["standard"], version = "^0.29.0"}
libgravatar = "^1.0.4"
bcrypt = "^4.1.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
